Add explicit cast
The length is never going to overflow int, but the compiler doesn't know
that.

Silence a warning here.

+++ plugins_tools/eid-viewer/verify.c
@@ -389,7 +389,7 @@ char* eid_vwr_x509_get_details(const void* certificate, size_t certlen) {
 	X509_print_ex(b, cert, 0, 0);
 	BIO_get_mem_ptr(b, &p);
 	rv = malloc(p->length + 1);
-	snprintf(rv, p->length + 1, "%.*s", p->length, p->data);
+	snprintf(rv, p->length + 1, "%.*s", (int)p->length, p->data);
 
 	BIO_free(b);
